+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Support `checkbutton` with specified `onvalue` and `offvalue` (when they do not have values 1 and 0)
 - Alias `root` `title` attribute as `text` & add children attribute to widgets
 - Alias `notebook` `frame` `:title` as `:text`
+- Alias message_box `:title` as `:text`
 
 ## 0.0.22